BackendEsp/imageDetection/requirements.txt
insightface
onnxruntime
opencv-python
numpy
easyocr
ultralytics
fastapi
uvicorn[standard]
openai
httpx
certifi
python-dotenv
